clean compile on Grower extending BaseSeller
NOTE that I left the Grower as a Grower and not a GrowerSeller. At this
point, it may make more sense to not go through every file and change
the Grower (including variables, function names, etc.) when there are
no GrowerUsers to get confused with. ALSO NOTE: it may make more sense
to change the current Handler object to HandlerUser and Trader to
TraderUser (and make the HandlerSeller a Handler)…may be easier to
distinguish moving forward.